// src/tasks/type/task.type.ts
export enum Status {
  LOCKED = 'LOCKED',
  AVAILABLE = 'AVAILABLE',
  CLAIMABLE = 'CLAIMABLE', // VERIFY
  COMPLETED = 'COMPLETED',
  PENDING = 'PENDING',
}

export enum Type {
  ONE_TIME = 'ONE_TIME',
  DAILY_STREAK = 'DAILY_STREAK',
}

export enum ReachVolumeType {
  TRADE = 'TRADE',
  DEPOSIT = 'DEPOSIT',
}

export enum Group {
  DAILY_CHECK_IN = 'DAILY_CHECK_IN',
  TRADE2AIRDROP = 'TRADE2AIRDROP',
  OUTBOND_MISSION = 'OUTBOND_MISSION',
  PARTNERSHIP = 'PARTNERSHIP',
}

export enum Condition {
  MAKE_A_TRADE = 'MAKE_A_TRADE',
  MAKE_A_DEPOSIT = 'MAKE_A_DEPOSIT',
  REACH_TRADE_VOLUME = 'REACH_TRADE_VOLUME',
  REACH_DEPOSIT_VOLUME = 'REACH_DEPOSIT_VOLUME',
  COMPLETE_CHILD_MISSION = 'COMPLETE_CHILD_MISSION',
  INVITE_FRIEND = 'INVITE_FRIEND',
  DAILY_TRADING = 'DAILY_TRADING',
  DAILY_CHECK_IN = 'DAILY_CHECK_IN',
  AFFILIATE_CLICK = 'AFFILIATE_CLICK',
  SUBSCRIBE_TWITTER = 'SUBSCRIBE_TWITTER',
  JOIN_TELEGRAM_GROUP = 'JOIN_TELEGRAM_GROUP',
  SUBSCRIBE_TELEGRAM_CHANNEL = 'SUBSCRIBE_TELEGRAM_CHANNEL',
}

export enum OtherTaskCode {
  SHARE_PNL_X = 'X001',
}